requirements.txt
langchain
langchain_community
langgraph
IPython
psycopg2
langchain-openai
python-dotenv
faiss-cpu
langchain-huggingface
pandas
ipykernel
langchain-groq
pydantic
FlagEmbedding
langchain_anthropic
